requirements.txt
django-simple-history
djangorestframework
Pillow
django-cors-headers
django-automatic-crud
djangorestframework-simplejwt
drf-yasg
setuptools
redis[hiredis]
celery
django-celery-results
eventlet
django-use-email-as-username
pylint
faker
gunicorn==20.0.4

opentelemetry-exporter-jaeger
opentelemetry-distro

opentelemetry-api 
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-django
opentelemetry-instrumentation-celery

pymysql
websockets
python-socketio
python-socketio[client]

channels==3.0.4
daphne==3.0.2
Django==4.0
